// src/app/model/Person.ts
export class Person {

    id?: number; 
    name: string;
    surname: string;
    about:string;
    ocupation: string;
    ubication: string;
    photo: string;
    banner: string;
    github: string;
    linkedin: string;
    whatsapp: string;
    email: string;
    user: string;
    password: string;
    
    constructor(){
        this.name = "";
        this.surname = "";
        this.about = "";
        this.ocupation = "";
        this.ubication = "";
        this.photo = "";
        this.banner = "";
        this.github = "";
        this.linkedin = "";
        this.whatsapp = "";
        this.email = "";
        this.user = "";
        this.password = "";
    }
    
}